+++ firebase-vscode/src/data-connect/execution/json-output.ts
@@ -1,0 +1,8 @@
+import vscode, {
+} from "vscode";
+
+export async function displayJsonFile(content: string) {
+  const doc = await vscode.workspace.openTextDocument(
+    {language: "json", content}
+  );
+}
